Adds note about editing food preferences and a shortcut to these settings. Also contains small formatting fixes, and centered the signup button.
Fixes #144 and #147 but feedback is needed if this is the solution we want to go with.

I am unsure about how the async thing works and wether ```goToSettings()``` should be that aswell..

Also; do we need to give a notice about changing food preference after the sign up period closed could be unnoticed by the Festivities Committee?

I tried to make the link scroll to the correct position (Food preferences visible) and maybe be higlighted, but I did not manage to figure out how to properly assign a ```GlobalKey()``` so that it gets a context for ```Scrollable.ensureVisible(key.currentContext)```..
I'll leave that for the future or somebody else.